refactoring type from Int to Long

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppslaunchpadauth/dto/PagedResult.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.hmppslaunchpadauth.dto
 data class PagedResult<T>(
   val page: Int,
   val exhausted: Boolean,
-  val totalElements: Int,
+  val totalElements: Long,
   val content: List<T>,
 )
